build: Remove androidx from sdk
Androidx imposes requirements on the App the implements it - they must use compileSdkVersion 28 and they must enable androidx and jetifier. I have moved androidx to the sample app only. The compileOnly dependency for androidx.annnotation will only be a dependency at compile time, so it won't affect Apps that implement the SDK.