fix: Use no_backup directory for caching
If an App has allowBackup=true and the user has backup enabled for their device, then the shared preferences and files directory will get backed up in the cloud. The saved files then get applied if they reinstall the app or install it on a new device. This causes problems because the public key is encrypted using a private key which is generated by the App. Also, this might be considered unexpected behavior by App developers and could be a security concern because data is being stored in a place that they did not intend. So I have changed the ConfigCache and the PublicKeyCache to use the no_backup directory which doesn't get saved when the device is backed up.